Split AstDelta
